Mesh generation from C# works, don't know how

+++ Hazard/src/Scripting/CSharp/Script/Rendering/BufferBindings.h
@@ -9,11 +9,11 @@ namespace Hazard::Scripting::CSharp::Bindings {
 		static BindMap GetBindings();
 	private:
 		static uint64_t VertexBuffer_Create_Native(void* createInfo);
-		static void VertexBuffer_Destroy_Native(uint32_t resourceID);
-		static uint32_t VertexBuffer_GetSize_Native(uint32_t resourceID);
+		static void VertexBuffer_Destroy_Native(uint64_t resourceID);
+		static uint32_t VertexBuffer_GetSize_Native(uint64_t resourceID);
 
 		static uint64_t IndexBuffer_Create_Native(void* createInfo);
-		static void IndexBuffer_Destroy_Native(uint32_t resourceID);
-		static uint32_t IndexBuffer_GetSize_Native(uint32_t resourceID);
+		static void IndexBuffer_Destroy_Native(uint64_t resourceID);
+		static uint32_t IndexBuffer_GetSize_Native(uint64_t resourceID);
 	};
 }
